+++ Diamond_backend_2/src/main/java/com/example/diamondstore/dto/ProductDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
@@ -11,6 +13,6 @@ public class ProductDTO {
     private String productName;
     private String description;
     private int mountId;
-    private float laborFee;
+    private BigDecimal laborFee;
     private String status;
 }
